Move App into new component
