+++ src/components/Home/Home.scss
