+++ src/components/Home/actionTypes.js
