+++ src/components/Home/container.js
@@ -1,5 +1,5 @@
 import {connect} from 'react-redux';
-import App from './App';
+import Home from './Home';
 import {setSendWithPhoto} from './actions';
 import {loadStoreList} from './../StoreList/actions';
 
@@ -11,4 +11,4 @@ const mapDispatchToProps = {
 export default connect(
   null,
   mapDispatchToProps
-)(App);
+)(Home);
